refactor apiDemos -> apidemos
