+++ jdi-light-mobile-tests/src/main/java/nativeapp/android/apidemos/ApiDemosApp.java
@@ -1,23 +1,23 @@
-package nativeapp.android.apiDemos;
+package nativeapp.android.apidemos;
 
 import com.epam.jdi.light.mobile.elements.pageobjects.annotations.JApp;
-import nativeapp.android.apiDemos.animation.AnimationPage;
-import nativeapp.android.apiDemos.animation.CheckboxesPage;
-import nativeapp.android.apiDemos.app.ActionBarDisplayOptionsPage;
-import nativeapp.android.apiDemos.app.ActionBarPage;
-import nativeapp.android.apiDemos.app.ActionBarUsagePage;
-import nativeapp.android.apiDemos.app.AppPage;
-import nativeapp.android.apiDemos.app.NotificationPage;
-import nativeapp.android.apiDemos.app.StatusBarPage;
-import nativeapp.android.apiDemos.views.ButtonsPage;
-import nativeapp.android.apiDemos.views.RatingBarPage;
-import nativeapp.android.apiDemos.views.SearchViewActionBarPage;
-import nativeapp.android.apiDemos.views.SearchViewAlwaysExpandedPage;
-import nativeapp.android.apiDemos.views.SearchViewFilterPage;
-import nativeapp.android.apiDemos.views.SearchViewPage;
-import nativeapp.android.apiDemos.views.SeekBarPage;
-import nativeapp.android.apiDemos.views.TextFieldsPage;
-import nativeapp.android.apiDemos.views.ViewsPage;
+import nativeapp.android.apidemos.animation.AnimationPage;
+import nativeapp.android.apidemos.animation.CheckboxesPage;
+import nativeapp.android.apidemos.app.ActionBarDisplayOptionsPage;
+import nativeapp.android.apidemos.app.ActionBarPage;
+import nativeapp.android.apidemos.app.ActionBarUsagePage;
+import nativeapp.android.apidemos.app.AppPage;
+import nativeapp.android.apidemos.app.NotificationPage;
+import nativeapp.android.apidemos.app.StatusBarPage;
+import nativeapp.android.apidemos.views.ButtonsPage;
+import nativeapp.android.apidemos.views.RatingBarPage;
+import nativeapp.android.apidemos.views.SearchViewActionBarPage;
+import nativeapp.android.apidemos.views.SearchViewAlwaysExpandedPage;
+import nativeapp.android.apidemos.views.SearchViewFilterPage;
+import nativeapp.android.apidemos.views.SearchViewPage;
+import nativeapp.android.apidemos.views.SeekBarPage;
+import nativeapp.android.apidemos.views.TextFieldsPage;
+import nativeapp.android.apidemos.views.ViewsPage;
 
 @JApp(appPackage = "io.appium.android.apis", appActivity = "io.appium.android.apis.ApiDemos")
 public class ApiDemosApp {
