+++ jdi-light-mobile-tests/src/main/java/nativeapp/android/apidemos/IndexPage.java
@@ -1,7 +1,6 @@
-package nativeapp.android.apiDemos;
+package nativeapp.android.apidemos;
 
 import com.epam.jdi.light.mobile.elements.common.app.Button;
-import com.epam.jdi.light.mobile.elements.common.app.android.StatusBar;
 import com.epam.jdi.light.mobile.elements.pageobjects.annotations.MobileFindBy;
 
 public class IndexPage {
