+++ jdi-light-mobile-tests/src/main/java/nativeapp/android/apidemos/animation/CheckboxesPage.java
@@ -1,4 +1,4 @@
-package nativeapp.android.apiDemos.animation;
+package nativeapp.android.apidemos.animation;
 
 import com.epam.jdi.light.mobile.elements.common.app.android.Checkbox;
 import com.epam.jdi.light.mobile.elements.pageobjects.annotations.MobileFindBy;
