+++ jdi-light-mobile-tests/src/main/java/nativeapp/android/apidemos/app/ActionBarDisplayOptionsPage.java
@@ -1,4 +1,4 @@
-package nativeapp.android.apiDemos.app;
+package nativeapp.android.apidemos.app;
 
 import com.epam.jdi.light.mobile.elements.common.app.android.ActionBar;
 import com.epam.jdi.light.mobile.elements.pageobjects.annotations.MobileFindBy;
