+++ jdi-light-mobile-tests/src/test/java/nativeapp_android/ActivityTestsInit.java
@@ -2,7 +2,7 @@ package nativeapp_android;
 
 import com.epam.jdi.light.driver.WebDriverFactory;
 import com.epam.jdi.light.mobile.elements.common.AppManager;
-import nativeapp.android.apiDemos.ApiDemosApp;
+import nativeapp.android.apidemos.ApiDemosApp;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
