+++ jdi-light-mobile-tests/src/test/java/nativeapp_android/ApiDemosTestInit.java
@@ -1,9 +1,8 @@
 package nativeapp_android;
 
 import com.epam.jdi.light.driver.WebDriverFactory;
-import com.epam.jdi.light.elements.init.PageFactory;
 import com.epam.jdi.light.mobile.elements.common.AppManager;
-import nativeapp.android.apiDemos.ApiDemosApp;
+import nativeapp.android.apidemos.ApiDemosApp;
 import org.testng.annotations.*;
