+++ jdi-light-mobile-tests/src/test/java/nativeapp_android/tests/RatingBarTests.java
@@ -1,9 +1,9 @@
 package nativeapp_android.tests;
 
 import com.epam.jdi.light.mobile.elements.composite.MobileScreen;
-import nativeapp.android.apiDemos.IndexPage;
-import nativeapp.android.apiDemos.views.RatingBarPage;
-import nativeapp.android.apiDemos.views.ViewsPage;
+import nativeapp.android.apidemos.IndexPage;
+import nativeapp.android.apidemos.views.RatingBarPage;
+import nativeapp.android.apidemos.views.ViewsPage;
 import nativeapp_android.ApiDemosTestInit;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.BeforeMethod;
