+++ jdi-light-mobile-tests/src/test/java/nativeapp_android/tests/SearchViewTests.java
@@ -3,12 +3,12 @@ package nativeapp_android.tests;
 import com.epam.jdi.light.mobile.elements.common.MobileKeyboard;
 import com.epam.jdi.light.mobile.elements.composite.MobileScreen;
 import io.appium.java_client.android.nativekey.AndroidKey;
-import nativeapp.android.apiDemos.IndexPage;
-import nativeapp.android.apiDemos.views.SearchViewActionBarPage;
-import nativeapp.android.apiDemos.views.SearchViewAlwaysExpandedPage;
-import nativeapp.android.apiDemos.views.SearchViewFilterPage;
-import nativeapp.android.apiDemos.views.SearchViewPage;
-import nativeapp.android.apiDemos.views.ViewsPage;
+import nativeapp.android.apidemos.IndexPage;
+import nativeapp.android.apidemos.views.SearchViewActionBarPage;
+import nativeapp.android.apidemos.views.SearchViewAlwaysExpandedPage;
+import nativeapp.android.apidemos.views.SearchViewFilterPage;
+import nativeapp.android.apidemos.views.SearchViewPage;
+import nativeapp.android.apidemos.views.ViewsPage;
 import nativeapp_android.ApiDemosTestInit;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
