+++ jdi-light-mobile-tests/src/test/java/nativeapp_android/tests/SeekBarTest.java
@@ -2,9 +2,9 @@ package nativeapp_android.tests;
 
 import com.epam.jdi.light.mobile.elements.common.app.Button;
 import com.epam.jdi.light.mobile.elements.composite.MobileScreen;
-import nativeapp.android.apiDemos.IndexPage;
-import nativeapp.android.apiDemos.views.SeekBarPage;
-import nativeapp.android.apiDemos.views.ViewsPage;
+import nativeapp.android.apidemos.IndexPage;
+import nativeapp.android.apidemos.views.SeekBarPage;
+import nativeapp.android.apidemos.views.ViewsPage;
 import nativeapp_android.ApiDemosTestInit;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
