+++ jdi-light-mobile-tests/src/test/java/nativeapp_android/tests/TextFieldTests.java
@@ -7,9 +7,9 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import static com.epam.jdi.light.settings.WebSettings.logger;
-import static nativeapp.android.apiDemos.IndexPage.*;
-import static nativeapp.android.apiDemos.views.TextFieldsPage.*;
-import static nativeapp.android.apiDemos.views.ViewsPage.textFieldsPage;
+import static nativeapp.android.apidemos.IndexPage.*;
+import static nativeapp.android.apidemos.views.TextFieldsPage.*;
+import static nativeapp.android.apidemos.views.ViewsPage.textFieldsPage;
 
 public class TextFieldTests extends ApiDemosTestInit {
